Fix floppy error reporting to allow booting from GIDE
Due to an error in the FDC seek command the boot routine always
thought to find a floppy disk so the hard disk boot was never
called.

The seek command now needs some time (which is for now a hardcoded
formula depending on the number of cylinders to move the head).