Add emulation of index hole pulse
This is required by the new D004 ROM (version 3.2) that
does not only check the ready line but also tests for
the presence of the index pulse to detect if a floppy
drive is present.

The current code will always generate the index pulse
as soon as the drive is selected (by out(f6h) in D004).