Fix generation of index hole pulse
Make sure the callback is only started if it is not already
running. Otherwise we get multiple callbacks as the device
select is sent in all cases before the drive is accessed.

Also make the initial delay till generating the signal a
bit shorter as the D004 ROM v2.0 code seems to get stuck
from time to time with the longer delay.