Change handling of reti opcode in peripheral chips
The reason for this change was the emulation of the
VDIP USB host controller module. The module drives
the PIO strobe line when it wants to send a character
to the cpu.
The problem was that a new irq arriving while the
previous irq handler is still running (e.g. due to
screen scrolling) was lost.
This happened in the because the reassert of the irq
was done in reti before the internal status was
cleared.