Change handling of the D004 terminal count signal
The fix to the index hole pulse generation that changed
the timing (commit 28b1233d1239af33baa1771ef7b050592b893958)
seems to have only shifted the odds for the problem to
show up.
Further testing points to the terminal count signal that
is sent after retrieving the data from the FDC. Up to
now only the output value of 0xc0 did have any effect on
the FDC state. Now every value that is not 0x00 will cause
the FDC to enter the result state.