Fix build on OpenSuSE 11.1
SuSE seems to ship with a different version of the netpbm
tools that are used to create the windows icons. The debian
version of ppmtowinicon accepts both -andppms and -andpgms
(but documents only -andpgms in the manpage). On SuSE the
previously used -andppms fails.
Switching to -andpgms should make both systems happy while
producing still the same output.