Fix kc85/3 caos version 3.1
SourceForge-Bug [ 2464849 ] Cave (Game) pointed out
that CAVE is not running correctly in the emulator.
Tracking down the problem led to wrong initialization
of the window 0 vector after power-on.
The reason is that there is a single byte error (the
byte at address fc70h was 00h but should be b7h) in
the caos rom that is included in the emulator. The
routine at this address is supposed to save the current
window vector at b79bh but due to the ROM error the
vector was loaded from 009bh.