Fix EOT handling in READ_DATA and WRITE_DATA command.
Normally the end of data transfer is signaled by the terminal
count pulse. If there is no terminal count signaled, automatically
the next sector will be read until the end of track is reached.
Checking for end of track works by comparing the currently read
sector with the EOT value in the read command. So if we now read
the last byte of the sector that is defined to be EOT the read
operation is aborted and the "abnormal termination" flag is set.