Handle TH_PF_444 and TH_PF_420 pixel formats
The older theora codes do not seem to handle TH_PF_444. Even
the decoder included in firefox (currently on debian testing)
does not show files encoded in this format correctly.
So for now we encode only quality >= 85% with the new and more
accurate pixel format. Everything below this quality will be
encoded in TH_PF_420 format.